Renamed: Effect => EffectConstructor which makes more specific to what it is and no conflict with Effect from Cats Effect
