--- core/src/main/scala/jdksymlink/effect/Effect.scala
@@ -1,32 +1,0 @@
-package jdksymlink.effect
-
-import cats.Id
-import cats.effect.IO
-
-/**
- * @author Kevin Lee
- * @since 2020-01-22
- */
-trait Effect[F[_]] {
-  def effect[A](a: => A): F[A]
-}
-
-object Effect {
-
-  def apply[F[_] : Effect]: Effect[F] = implicitly[Effect[F]]
-
-  implicit val ioEffect: Effect[IO] = new Effect[IO] {
-    override def effect[A](a: => A): IO[A] = IO(a)
-  }
-
-  implicit val idNonEffect: Effect[Id] = new Effect[Id] {
-    override def effect[A](a: => A): Id[A] = a
-  }
-
-}
-
-trait Effectful[F[_]] {
-  protected def EF: Effect[F]
-
-  def effect[A](a: => A): F[A] = EF.effect(a)
-}
