+++ core/src/main/scala/jdksymlink/effect/package.scala
@@ -11,17 +11,20 @@ import jdksymlink.core.data.YesOrNo
  */
 package object effect {
 
-  def readLnF[F[_] : Effect]: F[String] = Effect[F].effect(scala.io.StdIn.readLine)
-  def putStrLnF[F[_] : Effect](value: String): F[Unit] = Effect[F].effect(println(value))
+  def readLnF[F[_] : EffectConstructor]: F[String] =
+    EffectConstructor[F].effect(scala.io.StdIn.readLine)
 
-  def readYesOrNoF[F[_] : Effect : Monad](prompt: String): F[YesOrNo] = for {
+  def putStrLnF[F[_] : EffectConstructor](value: String): F[Unit] =
+    EffectConstructor[F].effect(println(value))
+
+  def readYesOrNoF[F[_] : EffectConstructor : Monad](prompt: String): F[YesOrNo] = for {
     _ <- putStrLnF[F](prompt)
     answer <- readLnF[F]
     yesOrN <-  answer match {
       case "y" | "Y" =>
-        Effect[F].effect(YesOrNo.yes)
+        EffectConstructor[F].effect(YesOrNo.yes)
       case "n" | "N" =>
-        Effect[F].effect(YesOrNo.no)
+        EffectConstructor[F].effect(YesOrNo.no)
       case _ =>
         readYesOrNoF[F](prompt)
     }
